requirements.txt
# Core LLM and orchestration
langchain==0.1.0
langchain-groq==0.0.1
langchain-community==0.0.10

# Vector database
chromadb==0.4.22

# Embeddings
sentence-transformers==2.3.1

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Optional but recommended
jupyter==1.0.0
pandas==2.1.4
numpy==1.26.3